REFERENCE: QGA-8   DESCRIPTION: Rename usersRoutes
