+++ server.js
@@ -5,7 +5,7 @@ import defaultErrorHandler from './middlewares/defaultErrorHandler';
 import mongoose from 'mongoose';
 import bodyParser from 'body-parser';
 import index from './routes/index';
-import user from './routes/user';
+import usersRoutes from './routes/usersRoutes';
 
 const app = express();
 const logger = require('./utils/logger')('server');
@@ -26,7 +26,7 @@ app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: true }));
 
 // Routing
-app.use(`/api/v${process.env.API_VERSION}/users`, user);
+app.use(`/api/v${process.env.API_VERSION}/users`, usersRoutes);
 app.use(`/api/v${process.env.API_VERSION}`, index);
 
 app.use('/uploads', express.static('uploads'));
